requirements.txt
elasticsearch==2.3.0
PyYAML==3.11
urllib3==1.14
wheel==0.24.0
